extract button insertion to a util
