+++ src/features/quick_tags.js
@@ -1,5 +1,4 @@
-import { cloneControlButton, createControlButtonTemplate } from '../utils/control_buttons.js';
-import { keyToCss } from '../utils/css_map.js';
+import { cloneControlButton, createControlButtonTemplate, insertControlButtonEditable } from '../utils/control_buttons.js';
 import { dom } from '../utils/dom.js';
 import { filterPostElements, getTimelineItemWrapper, postSelector } from '../utils/interface.js';
 import { megaEdit } from '../utils/mega_editor.js';
@@ -15,7 +14,6 @@ const symbolId = 'ri-price-tag-3-line';
 const buttonClass = 'xkit-quick-tags-button';
 const excludeClass = 'xkit-quick-tags-done';
 const tagsClass = 'xkit-quick-tags-tags';
-const controlIconSelector = keyToCss('controlIcon');
 
 let originalPostTag;
 let answerTag;
@@ -214,16 +212,8 @@ const processPosts = postElements => filterPostElements(postElements).forEach(po
   const tags = editedTagsMap.get(getTimelineItemWrapper(postElement));
   tags && addFakeTagsToFooter(postElement, tags);
 
-  const existingButton = postElement.querySelector(`.${buttonClass}`);
-  if (existingButton !== null) { return; }
-
-  const editIcon = postElement.querySelector(`footer ${controlIconSelector} a[href*="/edit/"] use[href="#managed-icon__edit"]`);
-  if (!editIcon) { return; }
-  const editButton = editIcon.closest('a');
-
   const clonedControlButton = cloneControlButton(controlButtonTemplate, { click: togglePopupDisplay });
-  const controlIcon = editButton.closest(controlIconSelector);
-  controlIcon.before(clonedControlButton);
+  insertControlButtonEditable(postElement, clonedControlButton, buttonClass);
 });
 
 popupElement.addEventListener('click', processBundleClick);
